requirements.txt
alpaca-trade-api==1.2.3
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
fastapi==0.68.0
h11==0.12.0
idna==3.2
msgpack==1.0.2
numpy==1.21.1
pandas==1.3.1
polygon-api-client==0.2.11
pydantic==1.8.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
websocket-client==1.1.0
websockets==9.1
